add optimizer and cost class
