+++ .vscode/settings.json
@@ -8,6 +8,7 @@
         "-p",
         "test_*.py"
     ],
+    "python.formatting.blackPath": "/Users/macvp/miniconda3/bin/black",
     "python.testing.pytestEnabled": false,
     "python.testing.unittestEnabled": true
 }
