+++ optimizer/cost.py
@@ -1,0 +1,22 @@
+
+"""Module docstrings."""
+
+import numpy as np
+
+from .cost import Model, Metric
+
+
+class Cost:
+    """Create a cost function from the encoding and the metric."""
+    def __init__(self, fn: np.ndarray, model: Model, metric: Metric):
+        self.model = model
+        self.metric = metric
+        self.fn = fn
+
+    def __call__(self, params) -> float:
+        fn_apprx = self.model(params)
+        return self.metric(fn_apprx - self.fn)
+
+    def grad(self, params) -> np.ndarray:
+        fn_apprx = self.model(params)
+        return self.metric.grad(fn_apprx - self.fn)
